Rewrite Zile in Vala
The test suite and make dist-check pass, but the editor has not been
otherwise much tested. The code still mostly looks like C; this is
deliberate for this commit, and will be improved in later commits.

There are still a few C files, implementing the Estr and Astr types, plus
the memrmem function. These will also be rewritten in Vala.