Translate estr into Vala
There is no more C source code in Zile proper, so simplify the build system.
We still need to detect a C compiler to build libgnu and memrmem, but do not
need compiler warnings etc.

Put estr into a convenience library, as Automake cannot generate build rules
for Vala to build the same Vala source file into two C files, so to avoid
conflicting rules it must be built only once and linked into zile and
estr-test.

Compiling Vala source with automake requires GNU Make, so check for it;
however, it is not required by the end user, and neither is valac, so print
only warnings if they are not found.

Turn leak detection back on.